Extend UIAlertController instead of defining new class for error alerts
